chore: move things top the routes where they are actually used
