+++ apps/token/src/components/connect-vega-wallet/index.ts
@@ -1,0 +1,1 @@
+export { ConnectToVega } from './connect-to-vega';
