+++ apps/token/src/routes/staking/home/index.ts
@@ -1,0 +1,1 @@
+export { Staking as default } from './staking';
