+++ apps/token/src/routes/staking/home/node-list.tsx
@@ -8,12 +8,12 @@ import {
 import type { AgGridReact } from 'ag-grid-react';
 import { useNavigate } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
-import { EpochCountdown } from '../../components/epoch-countdown';
-import { BigNumber } from '../../lib/bignumber';
+import { EpochCountdown } from '../../../components/epoch-countdown';
+import { BigNumber } from '../../../lib/bignumber';
 import { formatNumber } from '@vegaprotocol/react-helpers';
 import { ValidatorStatus } from '@vegaprotocol/types';
-import type { Nodes } from './__generated__/Nodes';
-import type { Staking_epoch } from './__generated__/Staking';
+import type { Nodes } from '../__generated__/Nodes';
+import type { Staking_epoch } from '../__generated__/Staking';
 import type { ColDef } from 'ag-grid-community';
 
 const VALIDATOR = 'validator';
