+++ apps/token/src/routes/staking/node/pending-stake.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import * as Sentry from '@sentry/react';
 import { Button, Callout, Intent, Loader } from '@vegaprotocol/ui-toolkit';
 import { useTranslation } from 'react-i18next';
-import { useAppState } from '../../contexts/app-state/app-state-context';
-import { BigNumber } from '../../lib/bignumber';
-import { removeDecimal } from '../../lib/decimals';
+import { useAppState } from '../../../contexts/app-state/app-state-context';
+import { BigNumber } from '../../../lib/bignumber';
+import { removeDecimal } from '../../../lib/decimals';
 import type { UndelegateSubmissionBody } from '@vegaprotocol/wallet';
 import { useVegaWallet } from '@vegaprotocol/wallet';
