+++ apps/token/src/routes/staking/node/stake-failure.tsx
