+++ apps/token/src/routes/staking/node/stake-success.tsx
@@ -1,7 +1,7 @@
 import { Callout, Intent } from '@vegaprotocol/ui-toolkit';
 import { useTranslation } from 'react-i18next';
 import { Link } from 'react-router-dom';
-import Routes from '../routes';
+import Routes from '../../routes';
 import type { StakeAction } from './staking-form';
 import { Actions, RemoveType } from './staking-form';
